learning/MongoDB_FastAPI/mongo_fastapi/pyproject.toml
[project]
name = "mongo_fastapi"
version = "0.1.0"
description = ""
authors = [
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"

]

[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
